Support for display
The display, which we use is the one from Nokia 5110. In this commit I've added suport for the display and a basic graphic interface. From now on the display will show preparation for menu, demo speed, demo distance, demo lap time. In the future I'm going to add suport for 2 other parameters to be shown and put icons for the menu.